feat: migrate :feature:detail-station to Compose Multiplatform
