+++ feature/detail-station/build.gradle.kts
@@ -1,34 +1,66 @@
+@file:OptIn(
+    org.jetbrains.compose.ExperimentalComposeLibrary::class,
+    org.jetbrains.kotlin.gradle.ExperimentalKotlinGradlePluginApi::class,
+)
+
 plugins {
-    alias(libs.plugins.gasguru.android.library)
+    alias(libs.plugins.gasguru.kmp.compose.library)
     alias(libs.plugins.gasguru.koin)
-    alias(libs.plugins.gasguru.compose.library)
     alias(libs.plugins.gasguru.proguard)
     alias(libs.plugins.stability.analyzer)
 }
 
+compose.resources {
+    publicResClass = true
+    packageOfResClass = "com.gasguru.feature.detail_station.generated.resources"
+}
+
+kotlin {
+    androidTarget {
+        instrumentedTestVariant.sourceSetTree.set(org.jetbrains.kotlin.gradle.plugin.KotlinSourceSetTree.test)
+    }
+
+    sourceSets {
+        commonMain.dependencies {
+            implementation(projects.core.analytics)
+            implementation(projects.core.ui)
+            implementation(projects.core.domain)
+            implementation(projects.core.model)
+            implementation(projects.core.uikit)
+            implementation(projects.core.common)
+            implementation(projects.navigation)
+            implementation(libs.coil3.compose)
+            implementation(libs.coil3.network.ktor3)
+            implementation(libs.kotlinx.datetime)
+            implementation(compose.components.uiToolingPreview)
+        }
+        commonTest.dependencies {
+            implementation(kotlin("test"))
+            implementation(libs.kotlinx.coroutines.test)
+            implementation(libs.turbine)
+            implementation(projects.core.testing)
+            implementation(libs.junit5.api)
+            implementation(libs.junit5.extensions)
+        }
+        androidUnitTest.dependencies {
+            implementation(libs.junit5.engine)
+        }
+    }
+}
+
 android {
     namespace = "com.gasguru.feature.detail_station"
+    defaultConfig {
+        testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
+    }
 }
 
 dependencies {
-    implementation(projects.core.analytics)
-    implementation(projects.core.ui)
-    implementation(projects.core.domain)
-    implementation(projects.core.model)
-    implementation(projects.core.uikit)
-    implementation(projects.core.common)
-    implementation(projects.navigation)
-    implementation(libs.koin.androidx.compose)
-    implementation(libs.constraint.layout)
-
-    testImplementation(projects.core.testing)
-    testImplementation(libs.junit5.api)
-    testImplementation(libs.junit5.extensions)
-    testImplementation(libs.kotlinx.coroutines.test)
-    testRuntimeOnly(libs.junit5.engine)
-
-    androidTestImplementation(projects.core.testing)
-    androidTestImplementation(projects.core.ui)
+    androidTestImplementation(libs.androidx.compose.ui.test.junit4)
+    debugImplementation(libs.androidx.compose.ui.test.manifest)
+    debugImplementation(compose.uiTooling)
     androidTestImplementation(libs.junit5.api)
     androidTestImplementation(libs.junit5.extensions)
+    androidTestImplementation(projects.core.testing)
+    androidTestImplementation(projects.core.ui)
 }
