+++ feature/detail-station/src/androidMain/AndroidManifest.xml
