+++ feature/detail-station/src/commonMain/composeResources/values/strings.xml
@@ -1,0 +1,36 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string name="go_station">Go to station</string>
+    <string name="fuels">Fuels</string>
+    <string name="open">Open</string>
+    <string name="close">Close</string>
+    <string name="price">Price</string>
+    <string name="direction">Direction</string>
+    <string name="schedule">Opening hours</string>
+    <string name="fuel_types">Fuel types</string>
+    <string name="station_detail">Details</string>
+    <string name="open_24h">24 hours</string>
+    <string name="monday_short">Monday</string>
+    <string name="tuesday_short">Tuesday</string>
+    <string name="wednesday_short">Wednesday</string>
+    <string name="thursday_short">Thursday</string>
+    <string name="friday_short">Friday</string>
+    <string name="saturday_short">Saturday</string>
+    <string name="sunday_short">Sunday</string>
+    <string name="update_now">Verified now</string>
+    <string name="update_minute_ago">Verified 1 minute ago</string>
+    <string name="update_minutes_ago">Verified %1$d minutes ago</string>
+    <string name="update_hour_ago">Verified 1 hour ago</string>
+    <string name="update_hours_ago">Verified %1$d hours ago</string>
+    <string name="update_day_ago">Verified 1 day ago</string>
+    <string name="update_days_ago">Verified %1$d days ago</string>
+    <string name="update_long_ago">Verified a long time ago</string>
+    <string name="navigate_with">Navigate with:</string>
+    <string name="share_with">Share with:</string>
+    <string name="share_fuel_prices">Fuel prices</string>
+    <string name="share_download">Download GasGuru and find the cheapest gas stations</string>
+    <string name="capacity_picker_title">Select capacity</string>
+    <string name="capacity_picker_range">Min 40 · Max 999</string>
+    <string name="capacity_picker_confirm">Confirm</string>
+    <string name="vehicle_default_name">My vehicle</string>
+</resources>
