+++ feature/detail-station/src/commonMain/kotlin/com/gasguru/feature/detail_station/platform/MapsNavigation.kt
@@ -1,0 +1,7 @@
+package com.gasguru.feature.detail_station.platform
+
+import androidx.compose.runtime.Composable
+import com.gasguru.core.model.data.LatLng
+
+@Composable
+expect fun rememberNavigateToMapsAction(): (LatLng) -> Unit
