+++ feature/detail-station/src/commonMain/kotlin/com/gasguru/feature/detail_station/platform/NotificationPermission.kt
@@ -1,0 +1,8 @@
+package com.gasguru.feature.detail_station.platform
+
+import androidx.compose.runtime.Composable
+
+@Composable
+expect fun rememberNotificationPermissionRequester(
+    onPermissionGranted: () -> Unit,
+): () -> Unit
