+++ feature/detail-station/src/commonMain/kotlin/com/gasguru/feature/detail_station/platform/ShareAction.kt
@@ -1,0 +1,6 @@
+package com.gasguru.feature.detail_station.platform
+
+import androidx.compose.runtime.Composable
+
+@Composable
+expect fun rememberShareAction(): (String) -> Unit
