+++ feature/detail-station/src/commonMain/kotlin/com/gasguru/feature/detail_station/ui/DetailStationEvent.kt
