--- feature/detail-station/src/main/res/drawable/ic_flag.xml
@@ -1,9 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="24dp"
-    android:height="24dp"
-    android:viewportWidth="960"
-    android:viewportHeight="960">
-  <path
-      android:pathData="M200,840v-680h360l16,80h224v400L520,640l-16,-80L280,560v280h-80Z"
-      android:fillColor="#000000"/>
-</vector>
