--- feature/detail-station/src/main/res/drawable/ic_fuel_station.xml
@@ -1,9 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="24dp"
-    android:height="24dp"
-    android:viewportWidth="960"
-    android:viewportHeight="960">
-  <path
-      android:pathData="M160,840v-640q0,-33 23.5,-56.5T240,120h240q33,0 56.5,23.5T560,200v280h40q33,0 56.5,23.5T680,560v180q0,17 11.5,28.5T720,780q17,0 28.5,-11.5T760,740v-288q-9,5 -19,6.5t-21,1.5q-42,0 -71,-29t-29,-71q0,-32 17.5,-57.5T684,266l-84,-84 42,-42 148,144q15,15 22.5,35t7.5,41v380q0,42 -29,71t-71,29q-42,0 -71,-29t-29,-71v-200h-60v300L160,840ZM240,400h240v-200L240,200v200ZM720,400q17,0 28.5,-11.5T760,360q0,-17 -11.5,-28.5T720,320q-17,0 -28.5,11.5T680,360q0,17 11.5,28.5T720,400ZM240,760h240v-280L240,480v280ZM480,760L240,760h240Z"
-      android:fillColor="#000000"/>
-</vector>
