--- feature/detail-station/src/main/res/values-es/strings.xml
@@ -1,36 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-    <string name="go_station">Ir a la estación</string>
-    <string name="fuels">Carburantes</string>
-    <string name="open">Abierto</string>
-    <string name="close">Cerrado</string>
-    <string name="price">Precio</string>
-    <string name="direction">Direccion</string>
-    <string name="schedule">Horario</string>
-    <string name="fuel_types">Tipos de combustible</string>
-    <string name="station_detail">Detalles</string>
-    <string name="open_24h">24 horas</string>
-    <string name="monday_short">Lunes</string>
-    <string name="tuesday_short">Martes</string>
-    <string name="wednesday_short">Miercoles</string>
-    <string name="thursday_short">Jueves</string>
-    <string name="friday_short">Viernes</string>
-    <string name="saturday_short">Sábado</string>
-    <string name="sunday_short">Domingo</string>
-    <string name="update_now">Verificados ahora</string>
-    <string name="update_minute_ago">Verificados hace 1 minuto</string>
-    <string name="update_minutes_ago">Verificados hace %d minutos</string>
-    <string name="update_hour_ago">Verificados hace 1 hora</string>
-    <string name="update_hours_ago">Verificados hace %d horas</string>
-    <string name="update_day_ago">Verificados hace 1 día</string>
-    <string name="update_days_ago">Verificados hace %d días</string>
-    <string name="update_long_ago">Verificados hace mucho tiempo</string>
-    <string name="navigate_with">Navegar con:</string>
-    <string name="share_with">Compartir con:</string>
-    <string name="share_fuel_prices">Precios de combustible</string>
-    <string name="share_download">Descarga GasGuru y encuentra las gasolineras más baratas</string>
-    <string name="capacity_picker_title">Selecciona la capacidad</string>
-    <string name="capacity_picker_range">Mínimo 1 · Máximo 999</string>
-    <string name="capacity_picker_confirm">Confirmar</string>
-    <string name="vehicle_default_name">Mi vehículo</string>
-</resources>
